refactor: restaurant 오탈자 수정
