+++ src/components/MainPage/Restaurant.ts
